Soooo much work went into those few lines in that file.
NOTES:
Its clear that there are many 'default' settings for css/html that
no one wants, and half of the troubleshooting is figuring out how to
undo the defaults.

Another issue is there alot of hack jobs people have come up with
that are obviously going to create technical debt in the future.